third_party/blink/web_tests/external/wpt/wai-aria/role/list-roles.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] orphan p with listitem role
  assert_equals: <p role="listitem" data-testname="orphan p with listitem role" data-expectedrole="paragraph" class="ex">x</p> expected "paragraph" but got "listitem"
[FAIL] orphan div with listitem role
  assert_false: Computed Role: "listitem" does not match any of the acceptable role strings in ["generic", "", "none"]: <div role="listitem" data-testname="orphan div with listitem role" class="ex-generic">x</div> expected false got true
Harness: the test ran to completion.
